Move `Namespace` object to `parsegen.utils`
