+++ parsegen/data.py
@@ -118,35 +118,4 @@ class Symbol(object):
 		
 		self.follow = self._union_set_with_values(self.follow, values)
 
-class Namespace(object):
-	"""Namespace
-	
-	Represents a read-only access to a dictionary through the . operator. Used
-	to allow easy access to options and suchlike.
-	"""
-	
-	def __init__(self, dict):
-		self.__dict__.update(dict)
-	
-	def __repr__(self):
-		return repr(self.__dict__)
-	
-	def __contains__(self, item):
-		return item in self.__dict__
-
-	# Not hashable
-	__hash__ = None
-	
-	def __cmp__(self, other):
-		return self.__neq__(other)
-	
-	def __eq__(self, other):
-		try:
-			return self.__dict__ == other.__dict__
-		except:
-			return False
-	
-	def __neq__(self, other):
-		return not self.__eq__(other)
-
