+++ test/test_data.py
@@ -140,52 +140,3 @@ class TestSymbol(object):
 		s.add_follow(['bar', 'baz'])
 		
 		assert s.follow == {'foo', 'bar', 'baz'}
-		
-class TestNamespace(object):
-	"""Test Namespace
-	
-	Tests the Namespacing object. This object is responsible for allowing easier
-	read_access to a dict.
-	"""
-	
-	def test_create(self):
-		
-		n = Namespace({})
-		assert n != None
-		
-		n = Namespace({"foo": "bar"})
-		assert n != None
-		assert n.foo == "bar"
-	
-	def test_invalid(self):
-		
-		assert_raises(TypeError, lambda : Namespace())
-	
-	def test_dict_access(self):
-		
-		d = {"foo": 14, "bar": 1243, "baz": "dsaf"}
-		
-		n = Namespace(d)
-		
-		for k, v in d.items():
-			assert getattr(n, k) == v
-	
-	def test_compare(self):
-		d = {"hello": "2134", "world": "rasdfa"}
-		
-		n1 = Namespace(d)
-		n2 = Namespace(d)
-		n3 = Namespace({})
-	
-		assert n1 == n2
-		assert not n1 != n2
-		assert n2 != n3
-		assert not n1 == n3
-	
-	def test_contains(self):
-		
-		n = Namespace({"foo": "", "bar": "string"})
-		
-		assert "foo" in n
-		assert "bar" in n
-		assert not "baz" in n
